+++ mdm/mdm/unmarshal_proto.go
@@ -302,6 +302,10 @@ func protoToSetting(s *mdmproto.Setting) Setting {
 		Item: s.GetItem(),
 	}
 	switch s.Item {
+	case "ApplicationConfiguration":
+		pbs := s.GetApplicationConfiguration()
+		setting.Identifier = nilIfEmptyString(pbs.GetIdentifier())
+		setting.ConfigurationData = pbs.GetConfigurationDictionaryData()
 	case "VoiceRoaming":
 		pbs := s.GetVoiceRoaming()
 		setting.Enabled = nilIfFalse(pbs.GetEnabled())