+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add API endpoint to retrieve APNS certificate (#503)
 * Remove deprecated `-apns` flags from server startup (#528)
 * Move API calls to list endpoints from HTTP GET to HTTP POST (#522, #523, #524, #525, #526)
+* Add support for the ApplicationConfiguration Setting (#521)
 
 ## [v1.4.0](https://github.com/micromdm/micromdm/compare/v1.3.1...v1.4.0) September 6 2018